Address review feedback on RandomStrategy NChooseK sampling
- sample_valid_nchoosek_features now takes seed: Optional[int] instead
  of a random.Random object, matching the convention used by
  Inputs.sample. The strategy passes self._get_seed().
- Drop the redundant feat.allow_zero = False reset before pinning a
  feature to bounds=[0, 0]; the ContinuousInput validator already
  exempts that case.
- Add explanatory comments around the NChooseK sampling branch in
  RandomStrategy._sample_with_nchooseks.
- Move the sample_valid_nchoosek_features tests from
  test_domain_nchoosek_combinatorics.py into test_domain.py and drop
  the redundant default-n test and the strategy scalability test.
